==> biomarin/bin/3.preprocessing_great.ipynb <==
from Bio.SeqIO.FastaIO import SimpleFastaParser
import numpy as np
import os, sys
import pandas as pd
import pybedtools as pb
import matplotlib.pyplot as plt
import seaborn as sns

LOCAL = True
if LOCAL is True:
    sys.path.append("/Users/sarahfong/tools/py_")
    PATH = "/Users/sarahfong/Desktop/local_data/Biomarin_Jun_2023/"
    MPRA = os.path.join(PATH, "delta_rank.csv")
    FASTA = os.path.join(PATH, "library_2", "Design", "biomarin-lib2-hg38-final.fasta")
    META_DATA= os.path.join(PATH, "lib2.meta_data.tsv")
    TILE_BED = os.path.join(PATH, "lib2.processed.bed")
    ENH_BED = os.path.join(PATH,'lib2.processed.merged.bed')
    GREAT_GENE = os.path.join(PATH, "GREAT", f"region2gene_lib2hg38.txt")
    GREAT_GENE_TSV = os.path.join(PATH, "GREAT", f"region2gene_lib2hg38.tsv")
                         
import config_readwrite as crw
import plot_params as pp
pp.fonts()

# evaluate GREAT

# get nearby coordinates using GREAT nearest 2 gene function

## load great annotations

locus_dict = {}
with open(GREAT_GENE, "r") as reader:
    for line in reader:
        if "#" not in line and line !="\n":
            
            locus, genes = line.strip("\n").split("\t")  # split on tab to get gene name

            info = [locus]
            
            for n, i in enumerate(genes.split(",")):
                if i != "NONE":
                    gene, dist = i.strip(" ").split(" ")
                    dist = dist.strip("()")
                    dist = dist.strip("+")
                else:
                    gene, dist = "na", 0
                
                locus_dict[f'{locus}-{n}'] = [locus, gene, int(dist)]


vstack = np.vstack(locus_dict.values())

great = pd.DataFrame(vstack)
great.columns = ["coor", "gene", 'dist']
great.head()

great.to_csv(GREAT_GENE_TSV, sep='\t', index=False)
