==> biomarin/bin/3_vierstra_merge_arch_redundancy.ipynb <==
Remove archetype prediction redundancies among archetypes that overlap MPRA library. 

Use sliding window and median archetype length to select archetype w best score. 

import os, sys

sys.path.append(os.getcwd()) # append current working directory

import config_readwrite as crw  # custom script for reading config
import pandas as pd
import pybedtools as pbt

import chr_functions

# config

config, cfn = crw.read("/wynton/home/ahituv/fongsl/biomarin/bin/config.neuron.mpra.orig.ini")

# read
section = "design"


LIB = "lib2"

if LIB == "lib1":
    TILES = config[section]["full_bed"]
    PATH = config[section]["PATH"]
    PREFIX = f"unq_tiles.{LIB}"
    VIERSTRA = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.bed")
    CLEANED = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.cleaned.bed")
    CLEANED_TABLE = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.cleaned.table.bed")


    section = "tfbs"
    crw.check(config, section)

    config[section]["vierstra"] = VIERSTRA  # write the tiles
    config[section]["vierstra.cleaned"] = CLEANED  # write the tiles
    config[section]["vierstra.cleaned.table"] = CLEANED_TABLE # max TFBS motif score, no redundancies within 13 bp. Not strand specific

else:
    
    TILES = "/wynton/group/ahituv/fongsl/projects/biomarin/library_2/design/biomarin-lib2-hg38-final.bed.gz"
    PATH = config[section]["path2"]
    PREFIX = "biomarin-lib2-hg38-final"
    VIERSTRA = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.bed.gz")
    CLEANED = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.cleaned.bed")
    CLEANED_TABLE = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.cleaned.table.bed")



    section = "tfbs2"
    crw.check(config, section)

    config[section]["vierstra"] = VIERSTRA  # write the tiles
    config[section]["vierstra.cleaned"] = CLEANED  # write the tiles
    config[section]["vierstra.cleaned.table"] = CLEANED_TABLE # max TFBS motif score, no redundancies within 13 bp. Not strand specific

crw.write(config, cfn)

# run bedtools intersection

# intersect coordinates with bed tools using the -wao command. 
cmd = " ".join(["qsub", 
                "$HOME/vierstra_archetype.sh", 
                TILES
                ])
if os.path.exists(VIERSTRA) is False:
    
    os.system(cmd)

# tiles 

if "#chr" not in (list(pd.read_csv(VIERSTRA, sep='\t', nrows=3))):
    
    # column names
    names = ["#chr", 'start_tile', "end_tile", "#chr_arch", "start_arch", "end_arch", "group",
        'arch', "strand", "model", "num_models", "overlap"]

    # load dataframe
    df = pd.read_csv(VIERSTRA, sep='\t', header=None, names=names, low_memory=False)

    # sort and save
    df.sort_values(by=list(df.columns[:3])).drop_duplicates().to_csv(VIERSTRA, sep='\t', index=False)
else:
    print('already rearranged columns')
    df = pd.read_csv(VIERSTRA, sep='\t', low_memory=False)

tf = df.loc[df["overlap"] > 0].copy()

# compute motif length
tf["motif_len"] = tf["end_arch"] - tf["start_arch"]

# annotate fully overlapping motifs
tf["motif_overlap"] = False

# full overlap when tf motif len == overlap len
tf.loc[tf['motif_len'] == tf["overlap"],
        "motif_overlap"] = True

tf.groupby("motif_overlap")["#chr"].count()  #about 10% false overlaps

## sanity check - each group, model has one motif length 
(not multiple motif lengths for a model, group)

- check the length(s) of the motifs. If more than one motif length per motif, should print
  
              model [1, 2, etc.]
- If one motif length, print

              model [1]

for col in ["model", "group"]:
    a = tf.groupby(col)["motif_len"].unique().reset_index()
    print(col,a["motif_len"].apply(lambda x: len(str(x).split(","))).unique()) # count the length of the unique motifs per model)
    
    print(len(set(tf[col])))

## prepare dataframe for sliding window

## sliding window function

def slideWindowCollapseGroup(df):
    print(df.shape)

    collection = {}  # collect results

    # cols to query
    cols = ["#chr_arch", "start_arch", "end_arch", "group", "strand", "arch"]

    # compute motif length
    df["len"] = df.end_arch-df.start_arch

    # compute median length
    median_len = df.len.describe()["50%"]

    # per chromosome, query starts
    for chr_ in set(df['#chr_arch']):
        print(chr_)
        test = df.loc[df['#chr_arch'] == chr_].copy()

        for start in test['start_arch'].unique():

            # test windows + median tfbs len range
            window = test.loc[test['start_arch'].between(
                start, int(start+median_len)), cols]

            # no group in window
            if window.shape[0] == 0:
                continue

            # if unique groups in window only
            elif len(set(window["group"])) == window.shape[0]:

                # add to collection
                collection[f'{chr_}:{start}'] = window

                continue

            # possibly non-unique
            else:
                # get max score among tfs
                # preserve strand information
                max_scores = window.groupby(["group", "strand"])[
                    'arch'].max().reset_index()

                # right join to keep only max_scores
                window = pd.merge(window, max_scores, how="right")

                collection[f'{chr_}:{start}'] = window[cols]

    # combine results
    cleaned = pd.concat(collection.values()).drop_duplicates()
    print("got rid of N redundacies=", df.shape[0]-cleaned.shape[0],
          "old:", df.shape,
          "new", cleaned.shape)

    return cleaned

## Sliding window to remove redundant TFBS archetype predictions

cleaned = slideWindowCollapseGroup(df)

## write cleaned file

cleaned.head()

cleaned.shape, df.shape

cleaned.to_csv(CLEANED, sep='\t', index=False)

# intersect tiles back into cleaned archetype dataset 

A = pbt.BedTool(TILES)
B = pbt.BedTool(CLEANED)

C = os.path.join(PATH, f"{PREFIX}.x.vierstra.archetypes.cleaned.re-intersect.lib1.bed")

A.intersect(B, wa=True, wb=True, output=C)

# format and pivot table for archetype data

if LIB == "lib1":
    keepcols =[0,1,2,5,10,14,15,16]
    names = ["#chr", 'start_tile', "end_tile","cell_origin","enh_id", "TFArch", "TFStrand", "TFscore"]
else:
    keepcols =[0,1,2,3,7,8,9]
    names = ["#chr", 'start_tile', "end_tile","id", "TFArch", "TFStrand", "TFscore"]
df = pd.read_csv(C, sep='\t', header=None, 
                usecols =keepcols, names=names
                ).drop_duplicates()
print(df.shape)

df.head()

## add tile coordinates back in

df = chr_functions.makeCoorAnnot(df, "#chr", 'start_tile', "end_tile", "tile")
df.head()

## groupby max score regardless of strand (creates redundancies. too hard.) and pivot

table = pd.pivot(df.groupby(["tile.coor", "TFArch"])["TFscore"].max().reset_index(), 
        index="tile.coor", 
         columns="TFArch", 
         values="TFscore"
        ).reset_index()

table.head()

## write table

table.drop_duplicates().to_csv(CLEANED_TABLE, sep='\t', index=False)

table.drop_duplicates().shape

CLEANED_TABLE

tabl
