==> biomarin/bin/4a.Activity_x_enh_loci_neighbor_support_bkgd.ipynb <==
20240529

sarahfong

# Goal

1. Systematically annotate whether loci have GABA/GLUT activity | Cell-type-specific activity support across multiple tiles.


    - "Support" means 3/5 adjacent tiles (60 bp within a 100bp window) meet any of the following criteria:
 
      
    - Cell-type-specific support
        - Cell-type-specific label activity (meaning that the tiles exceed the 95% shuffled PI in one cell type, but not another)
        - Significant FDR-corrected wilcoxon's P<0.05
        - Cell-type-specific activity difference exceeds the 95% CI of bootstrapped delta.mean (mean GABA - mean glut l2(RNA/DNA))
     
    -      
    -  Cell-type-active
        -3/5 adjacent tiles exceed the 95% PI of shuffled elements.
2. Assess how frequently loci are only active in 1 cell-type, both cell-types, and show subregional activity in loci that are active in both cell types.

3. Create background set of non-active and non-differentially active regions
  4. Annotate coordinates of cell-type-active, cell-type-specific elements

LOCAL = False


import numpy as np
import os, sys
import pandas as pd
import pybedtools as pb
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

import warnings
warnings.filterwarnings("ignore")


if LOCAL is True:
    sys.path.append("/Users/sarahfong/tools/py_")
    PATH = "/Users/sarahfong/Desktop/local_data/Biomarin_Jun_2023/"
    RE = os.path.join(PATH, "results")
    CONFIG_NAME="config.local.neuron.mpra.ini"
    RE_MAPS = os.path.join(RE, "enh_maps")
else:
    PATH="/wynton/group/ahituv/fongsl/projects/biomarin/data"
    RE = "/wynton/group/ahituv/fongsl/projects/biomarin/results"
    RE_MAPS = os.path.join(RE, "enh_maps")
    CONFIG_NAME="config.neuron.mpra.orig.ini"

# nomination dataframe
NOMS = os.path.join(PATH, "GABA_GLUT_DF.for.noms.tsv")
SUPPORT = os.path.join(PATH, "GABA_GLUT_DF.for.noms.tsv")

import config_readwrite as crw
import plot_params as pp
pp.fonts()

# config

config, cfn = crw.read(CONFIG_NAME)

section = 'lib2'
crw.check(config, section)

# read
DELTA_ACTIVITY_Z = config[section]["delta.mpra_centeredz"]

TILE_BED = config[section]["tile_bed"]
ENH_BED = config[section]["enh_bed"]

META_DATA = config[section]["metadata"]

SIG_ACTIVITY_Z = config[section]["sig-wilcoxon"]

# write
SUPPORT = os.path.join(RE, "support_neighbors.genomic_loci.tsv")
BKGD = os.path.join(RE, "bkgd.inactive.mpra.genomic_loci.tsv")
config[section]["support"] = SUPPORT
config[section]["bkgd"] = BKGD 
crw.write(config, cfn)

# Params

WINDOWSIZE, NSUPPORT = 5, 3

# load all the data

## library coordinates

lib = pd.read_csv(ENH_BED, sep='\t')
lib["enh_len"] = lib["end.enh"]-lib["start.enh"]
lib.head()

## MPRA activity summary

mpra = pd.read_csv(DELTA_ACTIVITY_Z, sep='\t')
mpra.head()

## MPRA reps activity

# get per rep activity
usecols = ['name',
            'gaba_1',
            'gaba_2',
            'gaba_3',
            'glut_1',
            'glut_2',
            'glut_3',
            'fdr_bool'
           ]
reps = pd.read_csv(SIG_ACTIVITY_Z, sep='\t', usecols=usecols)
reps.head()

## merge library enh and tile order w/ MPRA activity values

# merge library and MPRA activity, conserve library annotations (even when oligo doesn't have activity score)
lib_mpra = pd.merge(lib[["name", "enh.name", "tile.order", "enh_len"]], mpra, how="left") 
lib_mpra = pd.merge(lib_mpra, reps, how="left") # merge replicate measures, too

lib_mpra['sig_n_dif_n_bs'] = lib_mpra["sig"] + lib_mpra["celltype_dif"] + lib_mpra["bs"] # add combination of 3 sig measures

lib_mpra

# number of genomic enhancer loci that were tested for MPRA activity. The size of the full library is 1829 elements
drop_names = lib.loc[(lib["name"].str.contains("shuf")) |
             (lib["name"].str.contains("pos")) |
             (lib["name"].str.contains("neg")) |
             (lib["name"].str.contains("bkgd")), "name"]

genomic_lib = lib.loc[~lib["name"].isin(drop_names)].copy()
print(lib.shape, genomic_lib.shape)
genomic_lib[["enh.name"]].drop_duplicates().shape

# number of genomic enhancer loci that were tested for MPRA activity. The size of the full library is 1829 elements
genomic = lib_mpra.loc[(~lib_mpra["name"].isin(drop_names))].copy()
print(lib_mpra.shape, genomic.shape)
genomic[["enh.name"]].drop_duplicates().shape
genomic.tail()

# functions

def getSupportDif(enh_id, df):
    """count how many supports there is for differential activity"""

    # columns to keep
    t_cols = ["name", 'tile.order', "enh.name", "sig", "celltype_dif", "bs"]

    # subset df to enh_id w cols, drop duplicates, copy, and fill na w False
    t = df.loc[df["enh.name"] == enh_id,
               t_cols].drop_duplicates().copy().fillna(False)

    # turn bool into int, into str
    t[t_cols[-3:]] = t[t_cols[-3:]].astype(int)

    # str code the support col
    t["support_dif"] = t["sig"] + t["celltype_dif"] + t["bs"]

    # turn int into str
    t[t_cols[-3:]] = t[t_cols[-3:]].astype(str)

    t["support_dif_code"] = t["sig"] + t["celltype_dif"] + t["bs"]

    t["support_dif_name"] = None

    # name codes, so that they are legible
    name_codes = {None: "000",
                  "cat-only": "010",
                  "sig-only": "100",
                  "bs-only": '001',
                  "sig-n-cat": "110",
                  "sig-n-bs": "101",
                  "cat-n-bs": "011",
                  "all": "111"
                  }

    for name, code in name_codes.items():
        t.loc[t["support_dif_code"] == code, "support_dif_name"] = name

    return t[["name", "tile.order", "support_dif", "support_dif_code", "support_dif_name"
              ]]


def supportingDifNeighbors(enh_id, support, support_col, windowsize, nsupport):
    """Scan the nearest 5 windows and count support (number of significance tests the data passes)).
    return the tiles with the most significance from supported neighborhoods

    windowsize = 5  # define windowsize
    nsupport = 3  # number of significant tiles in window needed to call support
    """
    # list of the support for each tile (ordered) across enhancer
    support_list = list(support[support_col])

    # collect the tiles that have supporting neighbors, annotate max support neighbors
    support_neighborhoods, max_support_neighborhoods = [], []

    #  slide window across enhancer, finding supporting neighborhoods of activity
    for i in np.arange((support.shape[0] - windowsize)+1):

        # get neighborhood slice
        window_slice = support_list[i: i+windowsize]

        # count the number of significant measurements for each tile in window
        sig_array, sig_array_count = np.unique(
            window_slice, return_counts=True)

        # test whether window has 3+ tiles with *ANY* level of significance.
        # sig_array_count[0] counts the Falses
        if 0 not in sig_array or sig_array_count[0] < nsupport:
            
            # find the tile with the most evidence of significance
            m = window_slice.index(max(window_slice))
            max_support_neighborhoods.append(i+m)  # add that tile to the list
            support_neighborhoods.extend(list(np.arange(i, i+windowsize)))
            #print(i, window_slice, m, (i+m))

    support_neighborhoods = list(set(support_neighborhoods)) # make unique support neighbors into a list
    support_neighborhoods.sort()  # sort the list

    # collect info variables
    indexed_regions = {}
    first, last, total = 0,0,0

    # parse through support neighborhoods, indexing regions
    for i, tile in enumerate(support_neighborhoods):
    
        if i == 0:  # handle first tile
            first, last=tile, tile
    
        elif i == (len(support_neighborhoods)-1):  # handle the last tile
            
            region_id = enh_id + f".dif.{total}" 
            indexed_regions[region_id] = list(np.arange(first, tile+1))
            
        elif last+1 == tile: # continuous tiles, increase last count
            last = tile  # update last tile
        
        elif last+1 != tile: # noncontinueous tiles, annotate the last set of tiles, and reset to parse these tiles. 
            
            region_id = enh_id + f".dif.{total}" 
            indexed_regions[region_id] = list(np.arange(first, last+1))  
    
            # reset and update values for new region of tiles
            first, last=tile, tile
            total +=1

    return indexed_regions, set(max_support_neighborhoods)


def annotateSupportTiles(support_dict, supportDif, new_col):
    """annotate tiles with neighbor support list as bool"""
    
    supportdif[new_col] = 'None'

    for key, values in support_dict.items():
        for i in values: 
            supportdif.loc[supportdif["tile.order"] == i, new_col] = key
    
    return supportdif


def neighborActivity(longform_glut, longform_gaba, neighbor_support):
    glut_s = longform_glut.loc[longform_glut["tile.order"].isin(neighbor_support)].groupby(
        'tile.order')['Mean z-score ratio'].mean().reset_index()
    glut_s.rename(columns={"Mean z-score ratio": "glut.mean"}, inplace=True)

    gaba_s = longform_gaba.loc[longform_gaba["tile.order"].isin(neighbor_support)].groupby(
        'tile.order')['Mean z-score ratio'].mean().reset_index()
    gaba_s.rename(columns={"Mean z-score ratio": "gaba.mean"}, inplace=True)

    # print(pd.merge(glut_s, gaba_s ))
    # merge the mean activity of the gaba and glut tiles
    return pd.merge(glut_s, gaba_s)


def getSupportActive(enh_id, enh_df, cl, windowsize, nsupport):

    #enh_id, enh_df, cl, windowsize, nsupport = enhname, enhdf, cl, WINDOWSIZE, NSUPPORT
    """
    
    Atep1: count how many supports there is for differential activity in 5 window region
    cl = "gaba"
    df = lib_mpra
    enh_id =enhname
    """
    # columns to keep
    t_cols = ["name", 'tile.order', "enh.name", f"{cl}-label"]
    
    # subset df to enh_id w cols, drop duplicates, copy, and fill na w False
    support = enh_df.loc[enh_df["enh.name"] == enh_id,
                         t_cols].drop_duplicates().copy().fillna(False)
    
    support_col = f"{cl}-label"  # column to scan for consistent support
    
    """
    Index groups of regions with activity support. 
    
    Note: at worst, a window might have 2 tiles with no activity or no measurement of activity. 
    
        Step 1: Scan the nearest 5 windows and count support (number of significance tests the data passes)).
    return indexed subregions w/ tiles with the most significance from supported neighborhoods
        Step 2: index the regions with contiguous support. 
    """
    
    # list of the support for each tile (ordered) across enhancer
    support_list = list(support[support_col])
    
    support_neighborhoods = []  # collect the tiles that have supporting neighbors
    
    ###
    # Step 1: Scan 5 nearst windows and count support from label
    ###
    
    #  slide window across enhancer, finding supporting neighborhoods of activity
    for i in np.arange((support.shape[0] - windowsize+1)):
    
        # get neighborhood slice
        window_slice = support_list[i: i+windowsize]
    
        # count the number of significant measurements for each tile in window
        sig_array, sig_array_count = np.unique(
            window_slice, return_counts=True)
        
        if sig_array_count[0] < nsupport:  # count the falses
            support_neighborhoods.extend(list(np.arange(i, i+windowsize)))
          
    ###
    # Step 2: index support neighborhoods
    ###
    support_neighborhoods = list(set(support_neighborhoods)) # make unique support neighbors into a list
    support_neighborhoods.sort()  # sort the list

    # collect info variables
    indexed_regions = {}
    first, last, total = 0,0,0

    # parse through support neighborhoods, indexing regions
    for i, tile in enumerate(support_neighborhoods):
    
        if i == 0:  # handle first tile
            first, last=tile, tile
    
        elif i == (len(support_neighborhoods)-1):  # handle the last tile
            
            region_id = enh_id + f".{cl}.{total}" 
            indexed_regions[region_id] = list(np.arange(first, tile+1))
            
        elif last+1 == tile: # continuous tiles, increase last count
            last = tile  # update last tile
        
        elif last+1 != tile: # noncontinueous tiles, annotate the last set of tiles, and reset to parse these tiles. 
            
            region_id = enh_id + f".{cl}.{total}" 
            indexed_regions[region_id] = list(np.arange(first, last+1))  
    
            # reset and update values for new region of tiles
            first, last=tile, tile
            total +=1
    
    return indexed_regions
            
    

# main

# 1 parse support regions

support_dict = {}
enhnames=list(set(genomic["enh.name"].astype(str)))

#DEBUGGING: enhname="enh.208" | for enhname in np.random.choice(enhnames, size=3):
for enhname in enhnames:

    
    enhdf = genomic.loc[genomic["enh.name"] == enhname]
    
    # measure regions with significant different in activity
    supportdif = getSupportDif(enhname, enhdf) 
    
    # annotate differences in activity
    all_dif, max_dif = supportingDifNeighbors(enhname, supportdif, "support_dif", WINDOWSIZE, NSUPPORT)
    supportdif = annotateSupportTiles(all_dif, supportdif, "neighbor_dif") # annotate neighbor differences in activity
    
    
    # annotate regions with cell-type activity, does not have to be cell-type=specific
    for cl in ["gaba", "glut"]:
        supportactive = getSupportActive(enhname, enhdf, cl, WINDOWSIZE, NSUPPORT)
        supportdif = annotateSupportTiles(supportactive, supportdif, f"neighbor_{cl}_label") # annotate neighbor differences in activity
    
    supportdif["enh.name"] = enhname
    # add differences to dataset
    support_dict[enhname] = supportdif 

support = pd.concat(support_dict.values())
support.shape

support_dict['enh.424'].tail(10)

## 2. how many GLUT, GABA neighborhoods

# annotate if enhancer has support label
support['neighbor_either_label'] =support["neighbor_gaba_label"] + "."+ support["neighbor_glut_label"]
support.loc[support["neighbor_either_label"]!="None.None", "neighbor_either_label"] = "None" # re-write the none-none as none

print("either support N loci", len(support.loc[support["neighbor_either_label"]!="None", "enh.name"].unique()))

print("dif support N loci", len(support.loc[support["neighbor_dif"]!="None", "enh.name"].unique()))

print("gaba support N loci", len(support.loc[support["neighbor_gaba_label"]!="None", "enh.name"].unique()))
print("glut support N loci", len(support.loc[support["neighbor_glut_label"]!='None', "enh.name"].unique()))

774/1740, 665/1740, 187/1740, 606/1740 #total active enhancer loci, active gaba, active glut loci among all loci (active and inactive)  

187/665, 606/665  # total gaba, glut loci with activity among all active

support.to_csv(SUPPORT, sep='\t', index=False)

# collapse and index support

support=pd.read_csv(SUPPORT, sep='\t')

support.head()

support_ = pd.merge(lib[["name", "#chr", "start.tile", "end.tile"]], support).drop_duplicates()
support_.shape

# 3 background - regions w/o label for gaba, glut, or dif

inactive_bkgd = support_.loc[(support_["neighbor_gaba_label"].astype(str) == "nan") &
                             (support_["neighbor_glut_label"].astype(str) == "nan") &
                             (support_["neighbor_dif"].astype(str) == "nan")
                             ].drop_duplicates()

# rearrange columns to make bed-like file
inactive_bkgd = inactive_bkgd[[
    '#chr',
    'start.tile',
    'end.tile',
    'name',
    'tile.order',
    'support_dif',
    'support_dif_code',
    'support_dif_name',
    'neighbor_dif',
    'neighbor_gaba_label',
    'neighbor_glut_label',
    'enh.name',
    'neighbor_either_label']]

# change datatype
inactive_bkgd[['start.tile', 'end.tile']] = inactive_bkgd[[
    'start.tile', 'end.tile']].astype(int)
inactive_bkgd.head()

## merge background w pybedtools

bkgd = pb.BedTool.from_dataframe(inactive_bkgd)

merged_bkgd = bkgd.sort().merge(output=BKGD)  # merge and sort
bkgd = merged_bkgd.to_dataframe()  # make into dataframe
bkgd

## 4 assign support regions coordinates. 

for col in ["neighbor_either_label", "neighbor_gaba_label", "neighbor_glut_label", "neighbor_dif"]:

    # cols to keep 
    region_cols = ["#chr", col, "enh.name", "start.tile", "end.tile"]

    # keep only regions that have annotation (removes inactive regions)
    base =  support_.loc[support_[col].astype(str).str.contains("enh"), region_cols].copy()

    # get the min start for each active region
    region_starts =base.groupby(region_cols[:3])["start.tile"].min().reset_index()
    region_starts.rename(columns={"start.tile":"start.region"}, inplace=True)
    
    # get the end for each active region as the max start + 20bp (to account for the shift when the next tile loses activity)
    region_ends = base.groupby(region_cols[:3])["start.tile"].max().reset_index()
    region_ends.rename(columns={"start.tile":"end.region"}, inplace=True)
    region_ends['end.region'] = region_ends['end.region'] + 20

    # add start and ends of region together
    region = pd.merge(region_starts,region_ends).sort_values(by=['enh.name', "#chr", "start.region"])
    region[['start.region', 'end.region']] = region[['start.region', 'end.region']].astype(int)  # change datatype
    region=region[['#chr', 'start.region', 'end.region',  col, 'enh.name']]  # reorganize col order
    region["len"] = region["end.region"] - region["start.region"]  # compute length

    print(col, region["len"].describe())
    
    # write the region
    out = os.path.join(PATH, f"{col}.bed")

    region.to_csv(out, sep='\t', index=False)

## 2 (again) commonly active, cell-type-specific, and loci with distinct, cell-type-specific subregions.  

# get the loci where activity is labeld in both cell types
commonly_active_names = set(support.loc[(support['neighbor_either_label'].str.contains("gaba")) &
                                        (support['neighbor_either_label'].str.contains("glut")), "enh.name"].unique())

# get gaba labeled loci, can include commonly active names
gaba_active_names = set(support.loc[(support['neighbor_either_label'].str.contains("gaba"))
                                        , "enh.name"].unique())

# get glut labeled loci, can include commonly active names
glut_active_names = set(support.loc[(support['neighbor_either_label'].str.contains("glut")), "enh.name"].unique())

print(len(commonly_active_names), len(gaba_active_names), len(glut_active_names))

# GABA| GLUT - enhancer loci with commonly active activity, separate grammars, or gaba-only grammars
celltype_only, separate_grammars, shared_grammars =[], [], []

target_celltype="glut"


if target_celltype=="glut":
    activenames =glut_active_names
    ntarget= len(glut_active_names)
    other_celltype = "gaba"
else:
    activenames =gaba_active_names
    ntarget= len(gaba_active_names)
    other_celltype = "glut"


for i in activenames:
    t = support.loc[support["enh.name"]==i]
    if i in commonly_active_names:
        shared_grammars.append(i)
        
    elif t.loc[t[f"neighbor_{other_celltype}_label"]!="None"].shape[0] >0:  # see how often a glut active element is also in the locus
        separate_grammars.append(i)

    else: 
        celltype_only.append(i)


target_celltype, len(separate_grammars), len(celltype_only), len(shared_grammars), len(separate_grammars)/ngaba, len(celltype_only)/ngaba, len(shared_grammars)/ngaba

# case when we think a locus is commonly active, but also has differences in activity annotated. 
separate_and_common=[]
for i in commonly_active_names:
    t = support.loc[support["enh.name"]==i]
    if t.loc[t['neighbor_dif']!="None"].shape[0]>0:  # if there are differences in activity, annotated this locus as having both commonly activt and differently active regions 
        #print("separate grammars and common grammars")
        separate_and_common.append(i)
        

len(separate_and_common), (111-84)/111

128/665

19
