==> biomarin/bin/5_UCSC_GENOME_BROWSER_TRACKS.ipynb <==
LOCAL = False

import numpy as np
import os, sys
import pandas as pd

#import matplotlib.pyplot as plt
#import seaborn as sns
#from scipy import stats


import warnings
warnings.filterwarnings("ignore")


if LOCAL is True:
    sys.path.append("/Users/sarahfong/tools/py_")
    PATH = "/Users/sarahfong/Desktop/local_data/Biomarin_Jun_2023/"
    RE = os.path.join(PATH, "results")
    CONFIG_NAME="config.local.neuron.mpra.ini"
    RE_MAPS = os.path.join(RE, "enh_maps")
else:
    PATH="/wynton/group/ahituv/biomarin/data"
    RE = "/wynton/group/ahituv/fongsl/projects/biomarin/results"
    RE_MAPS = os.path.join(RE, "enh_maps")
    CONFIG_NAME="config.neuron.mpra.orig.ini"

# nomination dataframe
NOMS = os.path.join(PATH, "GABA_GLUT_DF.for.noms.tsv")

import config_readwrite as crw
import plot_params as pp
pp.fonts()

config, cfn = crw.read(CONFIG_NAME)

section = 'lib2'
crw.check(config, section)

RE = config["path"]["results_path"]
DELTA_ACTIVITY = config[section]["delta.mpra_centered"]
DELTA_ACTIVITY_Z = config[section]["delta.mpra_centeredz"]

FASTA = config[section]["fasta"]

TILE_BED = config[section]["tile_bed"]
ENH_BED = config[section]["enh_bed"]

GREAT_GENE_TSV = config[section]["great_nearest"]
META_DATA = config[section]["metadata"]

SIG_ACTIVITY_Z = config[section]["sig-wilcoxon"]
VISTA = config[section]["vista"]

SUPPORT = config[section]["support"] # neighboring tile support
TFBS = config["tfbs2"]["vierstra.motif.loc.nored"] # TFBS archetypes where score is in 95% of motif scores, binned by length, no redundants
# i.e. the strongest of the motif matches. 

LOSSES_AND_GAINS = config["tfbs2"]["vierstra.motif.loc.nored.loss.gain"]  # write the tiles with losses and gains only. 

# BEDS to write

TILE_BED_TRACK = os.path.join(RE, "track.tile.bed")
ENH_BED_TRACK = os.path.join(RE, "track.enh.bed")
TFBS_BED_TRACK = os.path.join(RE, "track.tfbs.bed")
GABA_BED_TRACK = os.path.join(RE, "track.gaba.MPRA.bed")
GLUT_BED_TRACK = os.path.join(RE, "track.glut.MPRA.bed")
#TFBS_BED_TRACK = os.path.join(RE, "track.tfbs.bed")

tile = pd.read_csv(ENH_BED, sep='\t').dropna()

# change datatype 
tile[["start.tile", "end.tile", "start.enh", "end.enh"]]=tile[["start.tile", "end.tile", "start.enh", "end.enh"]].astype(int)

tile_track_cols = ["#chr", "start.tile", "end.tile", "name"]
enh_track_cols = ["#chr", "start.enh", "end.enh", "enh.name"]

tile[tile_track_cols].drop_duplicates().to_csv(TILE_BED_TRACK, sep='\t', index=False)
tile[enh_track_cols].drop_duplicates().to_csv(ENH_BED_TRACK, sep='\t', index=False)

## TFBS

tf = pd.read_csv(TFBS, sep='\t').drop_duplicates()
tf.head()

tf_track_cols = ["#chr", "start_arch", "end_arch", "group", "arch", "strand", "start_arch", "end_arch", ]
tf[tf_track_cols].drop_duplicates().to_csv(TFBS_BED_TRACK, sep='\t', index=False)

tf[tf_track_cols].drop_duplicates()

# active

active = pd.read_csv(DELTA_ACTIVITY_Z, sep='\t')
active.head()

active = pd.merge(tile[tile_track_cols], active)

active_gaba_cols = tile_track_cols[::][:-1] + ["gaba"]
active_glut_cols = tile_track_cols[::][:-1] + ["glut"]

active[active_gaba_cols].drop_duplicates().to_csv(GABA_BED_TRACK, sep='\t', index=False, header=False)
active[active_glut_cols].drop_duplicates().to_csv(GLUT_BED_TRACK, sep='\t', index=False, header=False)

GABA_BED_TRACK

active[active_gaba_cols].drop_duplicates().sort_values(by=active_gaba_cols[:2])#.drop_duplicates()

active[active_gaba_cols].dtypes
