==> biomarin/bin/5d.merge.candidate.elements.celltype-specific.ipynb <==
from Bio.SeqIO.FastaIO import SimpleFastaParser
import numpy as np
import os, sys
import pandas as pd
import pybedtools as pb
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats
import warnings
warnings.filterwarnings("ignore")

LOCAL = True
if LOCAL is True:
    sys.path.append("/Users/sarahfong/tools/py_")
    PATH = "/Users/sarahfong/Desktop/local_data/Biomarin_Jun_2023/"
    RE = os.path.join(PATH, "results")
    CONFIG_NAME="config.local.neuron.mpra.ini"
    RE_MAPS = os.path.join(RE, "enh_maps")
else:
    PATH="/wynton/group/ahituv/biomarin/data"
    RE = "/wynton/group/ahituv/biomarin/results"
    RE_MAPS = os.path.join(RE, "enh_maps")
    CONFIG_NAME="config.neuron.mpra.orig.ini"

# nomination dataframe
NOMS = os.path.join(PATH, "GABA_GLUT_DF.for.noms.tsv")

# summarized across significant loci with support neighbor significance in 3/5 neighbors
CANDIDATE_REGIONS = os.path.join(PATH, "CANDIDATE.SIG.REGIONS.GABA.GLUT.tsv")
SUMMED_CANDIDATE_REGIONS = os.path.join(PATH, "SUMMARIZED.CANDIDATE.SIG.REGIONS.GABA.GLUT.tsv")

import config_readwrite as crw
import plot_params as pp
pp.fonts()

config, cfn = crw.read(CONFIG_NAME)

section = 'lib2'
crw.check(config, section)

if "fasta" not in list(config[section]):  # write files to config

    MPRA = os.path.join(PATH, "delta_rank.csv")

    FASTA = os.path.join(PATH, "library_2", "Design",
                         "biomarin-lib2-hg38-final.fasta")
    META_DATA = os.path.join(PATH, "lib2.meta_data.tsv")
    TILE_BED = os.path.join(PATH, "lib2.processed.bed")
    ENH_BED = os.path.join(PATH, 'lib2.processed.merged.bed')
    GREAT_GENE_TSV = os.path.join(PATH, "GREAT", f"region2gene_lib2hg38.tsv")
    DELTA_ACTIVITY = os.path.join(PATH, "deltaMPRA.gaba.minus.glut.tsv")
    DELTA_ACTIVITY_Z = os.path.join(
        PATH, "deltaMPRA.gaba.minus.glut.zscore.tsv")
    SIG_ACTIVITY_Z = os.path.join(PATH, "WilcoxonSig.gaba.glut.zscore.tsv")
    VISTA = os.path.join(PATH, "vista.lib2.tiles.bed")

    # delta MPRA activity
    config[section]["delta.mpra_richa"] = MPRA
    config[section]["delta.mpra_centered"] = DELTA_ACTIVITY
    config[section]["delta.mpra_centeredz"] = DELTA_ACTIVITY_Z

    config[section]["fasta"] = FASTA

    # genome coordinates
    config[section]["tile_bed"] = TILE_BED
    config[section]["enh_bed"] = ENH_BED

    # GREAT
    config[section]["great_nearest"] = GREAT_GENE_TSV

    # meta_data
    config[section]["metadata"] = META_DATA

    # significance GABA v. GLUT
    config[section]["sig-wilcoxon"] = SIG_ACTIVITY_Z

    # vista
    config[section]["vista"] = VISTA
    crw.write(config, cfn)

else:
    MPRA = config[section]["delta.mpra_richa"]
    DELTA_ACTIVITY = config[section]["delta.mpra_centered"]
    DELTA_ACTIVITY_Z = config[section]["delta.mpra_centeredz"]

    FASTA = config[section]["fasta"]

    TILE_BED = config[section]["tile_bed"]
    ENH_BED = config[section]["enh_bed"]

    GREAT_GENE_TSV = config[section]["great_nearest"]
    META_DATA = config[section]["metadata"]

    SIG_ACTIVITY_Z = config[section]["sig-wilcoxon"]
    VISTA = config[section]["vista"]
    

enh = pd.read_csv(NOMS, sep='\t')

# nominate candidates

## functions!

### get min dist to genes

def getMinDist(df):
    """return str of gene, min distance tuples to gene TSSs per enhancer. 
        Gene TSS annotations from GREAT. 
        note - most enhancers map to two genes, which means two TSSs
    """

    genes = []

    # per gene
    for gene in df["gene"].unique():
        genes.append(gene)
        print(gene)

        # get the gene dataframe
        g = df.loc[(df["gene"] == gene)].copy()

        # if upstream TSS, return min value
        g = g.loc[g["dist"] != "NONE"]
        if len(g) > 0:
            if int(g.dist.min()) > 1:
                genes.append(g.dist.min())

            # elif downstream TSS, return max negative value (min)
            elif int(g.dist.min()) < 1:
                genes.append(g.dist.max())
                
            elif list(set(g.dist)) == "na":
                print(gene,  "TO REMOVE")
                genes = genes[:-1]  # remove gene from list, no distance

    if len(genes) == 0:  # handle cases where no gene is mapped to element.
        genes = ["None"]

    new_genes = []

    for gene in genes:
        new_genes.append(str(gene))

    return "_".join(new_genes)

### N significant labels per tile

def getSupport(enh_id, df):
    """count how many supports there is for differential activity"""
    
    # columns to keep
    t_cols = ["name", 'tile.order', "enh.id", "sig", "celltype_dif", "bs"]
    
    # subset df to enh_id w cols, drop duplicates, copy, and fill na w False
    t = df.loc[df["enh.id"] == enh_id, t_cols].drop_duplicates().copy().fillna(False)
    
    # turn bool into int, into str
    t[t_cols[-3:]]=t[t_cols[-3:]].astype(int)
    
    # str code the support col
    t["support"] =  t["sig"] + t["celltype_dif"] + t["bs"]
    
    # turn int into str
    t[t_cols[-3:]]=t[t_cols[-3:]].astype(str)

    t["support_code"] =  t["sig"] + t["celltype_dif"] + t["bs"]    
    
    t["support_name"] = None
    
    # name codes, so that they are legible
    name_codes = {None:"000", 
                  "cat-only":"010", 
                  "sig-only":"100", 
                  "bs-only":'001',
                  "sig-n-cat":"110", 
                  "sig-n-bs":"101", 
                  "cat-n-bs":"011", 
                  "all": "111"
                 }
    
    for name, code in name_codes.items():
        t.loc[t["support_code"]==code, "support_name"] = name


    return t[["name", "tile.order", "support", "support_code", "support_name"
             ]]

### Neighbor analysis - N significant labels from 5-tile window

def supportingNeighbors(support):

    """Scan the nearest 5 windows and count support (number of significance tests the data passes)).
    return the tiles with the most significance from supported neighborhoods, return set of tiles that support each other
    """
    windowsize = 5  # define windowsize

    support_list = list(support["support"]) # list of the support for each tile (ordered) across enhancer 

    support_neighborhoods = []  # collect the tiles that have supporting neighbors
    all_neighbors = []
   
    #  slide window across enhancer, finding supporting neighborhoods of activity
    for i in np.arange((support.shape[0] - windowsize+1)):
        
        # get neighborhood slice
        window_slice = support_list[i: i+windowsize]

        # count the number of significant measurements for each tile in windown
        sig_array, sig_array_count = np.unique(window_slice, return_counts=True)
        
        # test whether window has 3+ tiles with *ANY* level of significance. 
        if 0 not in sig_array or sig_array_count[0]<3:
            
            # find the tile with the most evidence of significance
            m = window_slice.index(max(window_slice))
            support_neighborhoods.append(i+m)  # add that tile to the list
            #print("add", i, m,  "window", window_slice,  "tile", (i+m))
        
            # get all significant neighbors when sig neighborhood
            for n, slice in enumerate(window_slice):
                if slice> 0:
                    #print("add all neighbors", n +i)
                    all_neighbors.append((n+i))
                    
    
            
    #print(set(all_neighbors))

    return set(support_neighborhoods), set(all_neighbors)

def neighborActivity(longform_glut, longform_gaba, neighbor_support):
    glut_s = longform_glut.loc[longform_glut["tile.order"].isin(neighbor_support)].groupby('tile.order')['Mean z-score ratio'].mean().reset_index()
    glut_s.rename(columns={"Mean z-score ratio":"glut.mean"}, inplace=True)
    
    gaba_s = longform_gaba.loc[longform_gaba["tile.order"].isin(neighbor_support)].groupby('tile.order')['Mean z-score ratio'].mean().reset_index()
    gaba_s.rename(columns={"Mean z-score ratio":"gaba.mean"}, inplace=True)

    #print(pd.merge(glut_s, gaba_s ))
    # merge the mean activity of the gaba and glut tiles
    return pd.merge(glut_s, gaba_s )


def fillInSigGaps(sig_neighbor_tiles):
    """create blocks of tiles that are contiguously significant. Accept 1 gap"""
    
    
    gap_limit = 3 # anything this size or larger is excluded
    counter, start = 0, 0
    contig = []
    for i in sig_neighbor_tiles:
        
        if i == 0:  # if the list starts with zero
            continue
        elif i == sig_neighbor_tiles[0]:
            start, counter = i, i
            
        elif i == sig_neighbor_tiles[-1]:  # when we're at the end of the sig_tiles
            print("last one")
            if i > (counter + gap_limit):
             
                run = [i for i in np.arange(start, counter+1)]
            else:
                run = [i for i in np.arange(start, i+1)]
            contig.append(run)
            
        elif i == (counter + 1):  # continuous
            counter +=1
            run = [i for i in np.arange(start, counter+1)]
            
        elif i == (counter +gap_limit):  # gap of 1
            counter +=2
            run = [i for i in np.arange(start, counter+1)]
            
        elif i > (start+counter +2) & start !=0 : # gap > 1
            run = [i for i in np.arange(start, counter+1)]
            print("broken at", i, "range", start, 'to', counter, run)
            contig.append(run)
            start, counter = i, i
            
    print(contig)
        
    return contig

### filter for tiles with 3 significant labels and direction

def getSigTileIdDirection(direction, df, n_criteria=3):
    """return tile ids in direction which 
    (1) have significant activity differences
    (2) have categorical activity differences
    (3) exceed the bootstrapped 95 CIs from delta distribution
    """
    cols = ["bs",
            'direction',
            "name", "sig"]
    
    if n_criteria ==1:
        sig_col = "sig"
    elif n_criteria ==2:
        sig_col1 = "sig_n_dif"
        sig_col2 = "celltype_dif"
        sig_col3 = "bs"
        test = df.loc[
                  (df["sig"] == True)&
                    ((df[sig_col2] == True)|
                    (df[sig_col3] == True)),
                  cols].drop_duplicates().copy()
        sig_col = sig_col1 + sig_col2 + sig_col3
    else:
        sig_col = "sig_n_dif_n_bs"
        test = df.loc[
                  (df[sig_col] == True),
                  cols].drop_duplicates().copy()
   

    print(sig_col) 
    
    candidate_set = set(test.loc[test["direction"] == direction, "name"])

    return candidate_set

### make longform enhancer dataframe

def getLongFormEnhInfo(df, mpra_celltype):
    """ return dataframes for plotting gaba/glut longform tile activity and significance
    """
    # do some data formatting, sorting`
    df["tile.order"] = df["tile.order"].astype(int)
    df = df.sort_values(by="tile.order")

    # get dataframe of tile.order x significant differences
    sd = pd.melt(df, id_vars="tile.order",
                 value_vars="sig_n_dif_n_bs",
                 value_name="sig"
                 ).drop_duplicates().reset_index()
    sd["y"] = 3
    sd["sig"] = sd["sig"].replace(False, None)
    
    # new variable
    VAR_NAME = "Mean z-score ratio"

    # get replicates of gaba, glut and melt into longform
    if mpra_celltype == "gaba": 
        cols = ["gaba_1", "gaba_2", "gaba_3"]
        
    elif mpra_celltype =="glut":
        cols = ["glut_1", "glut_2", "glut_3"]

    # longform GABA
    longform = pd.melt(df, id_vars="tile.order",
                            value_vars=cols,
                            value_name=VAR_NAME).drop_duplicates().reset_index()

   
    # add back significance column
    longform = pd.merge(longform, sd, how="left", left_on="tile.order",
                             right_on='tile.order').drop(columns=["index_x", "index_y"])
    

    # add back missing tiles
    longform = addMissingTiles(longform, VAR_NAME)
    
    # change tile.order datatype
    longform["tile.order"]= longform["tile.order"].astype(int)
    
    # fill any missing data w zeros
    longform[VAR_NAME]=longform[VAR_NAME].fillna(0)
    
    return longform, VAR_NAME

### enh_id given tile_id

def getEnhid(tile_coor, df):
    """return enhancer coordinate for tile_id"""
    enh = set(df.loc[df["coor"] == tile_coor, 'enh.id'])

    if len(enh) > 1:
        print("more than one enh_id", enh)

    # return the first (and hopefully only) enh
    if len(enh)>0:
        return list(enh)[0]
    else:
        return None

### add back missing tiles

def addMissingTiles(longform, var_name):
    """add back missing tile values for a dataframe"""

    # make an empty dataframe
    empty = pd.DataFrame(
        np.arange(longform["tile.order"].min(),
                  longform["tile.order"].max()+1),
        columns=["tile.order"]
    )
    # add back empty, drop old index
    longform = pd.merge(empty, longform,
                             how="left")
    # fill na
    longform[var_name] = longform[var_name].fillna(0)

    return longform

### line plot for enhancer landscape

def plot_lineplot(longform_gaba, longform_glut, var_name, out, enh_id,  enh_name, gene_name, support, plot_support_annot):

    if len(set(longform_gaba['tile.order'])) > 20:
        fig, ax = plt.subplots(figsize=(12, 4))
    else:
        fig, ax = plt.subplots(figsize=(8, 4))

    # print max scores
    max_gaba, max_glut = longform_gaba[var_name].max(),\
    longform_glut[var_name].max()

    # min scores
    min_gaba, min_glut = longform_gaba[var_name].min(),\
    longform_glut[var_name].min()

    # get max of two for plotting purposes
    max_of_max = max(max_gaba, max_glut)
    min_of_min = min(min_gaba, min_glut)

    if max_of_max < 3.2:
        ymax = 3.2
        ymin = (ymax*-1) - 1
    else:
        ymax = max_of_max + 1
        ymin = (ymax*-1) - 1

    print('max gaba', max_gaba, "max_glut", max_glut)

    # plot longform gaba across enhancer
    sns.pointplot(x="tile.order", y=var_name, data=longform_gaba.fillna(0),
                  label="gaba",
                  join=False,
                  errorbar=("sd"),
                  color="orange", ax=ax)
    plt.setp(ax.collections, alpha=.3)  # alpha
    x="tile.order"
    data=longform_glut.fillna(0)

    # plot longform glut activity across enhancer
    sns.pointplot(x="tile.order", y=var_name, data=data,
                  label="glut",
                  join=False,
                  errorbar=("sd"),
                  color="blue", 
                 # ax=ax
                 )
    plt.setp(ax.collections, alpha=.3)  # alpha

    # where to place support for significance
    support["y"] = ymax + 0.5

    x = "tile.order"
    y = "y"
    data = support.fillna(0).reset_index(drop=True)
    

    if plot_support_annot == "support_name":
        hue = "support_name"  # "support"
        hue_order = ["all", "sig-n-bs", "sig-n-cat",
                     'cat-n-bs', "cat-only", "sig-only", "bs-only"]
        markers = ["*", 'o', "s", "p", ".", ".", "."]
        marker_title = "sig criteria"
        palette="tab20b"

    else:
        hue='support'
        hue_order = [3, 2, 1]
        markers = ['*', "o", "."]
        marker_title = "sig criteria N"
        palette="tab10"
    

    ax2 = ax.twiny()
    sns.pointplot(x=x, y=y, data=data, hue=hue,
                    hue_order=hue_order,
                    markers=markers,
                    palette=palette,
                    errorbar=None,
                    join=False, ax=ax2,
                    scale=1.5
                 )
    ax2.set_xticklabels("")

    # mark tiles w/ all three sig. 
    if True in set(longform_glut['sig']) or True in set(longform_gaba["sig"]):
        longform_glut["y"] = ymax 

        x="tile.order"
        y="y"
        hue="sig"
        data=longform_glut[[x,y,hue]].drop_duplicates().fillna(False)
    
        sns.pointplot(x=x, y=y, hue=hue, data=data, 
                      markers=[' ', "X"],
                      color="r",
                      errorbar=None,
                      join=False, ax=ax
                      )

    # zero line
    ax.axhline(0, c="grey", ls="--")

    # ax title, labels
    ax.set(title=f"{gene_name}bp to TSS\n{enh_id}",
           ylabel="mean z-score",
           ylim=(ymin, ymax + 1)
           )

    
    # Set major ticks for x axis
    ntiles = longform_gaba['tile.order'].max()
    if ntiles >= 100:
        segment = 10
    else:
        segment = 5

    major_xticks = np.arange(0, ntiles, segment)
    ax.set_xticks(major_xticks)

    #ax2.set_yticklabels("")
    
    # legend
    ax.legend(title='key', frameon=False,
              bbox_to_anchor=(0.75, -0.25))
    ax2.legend(title=marker_title, bbox_to_anchor=(1, -0.25), frameon=False)

    # save figure
    plt.savefig(out, bbox_inches="tight")

    plt.show()
    plt.close()
    

### tile landscape

def tileLandscape(candidate_list, df, top_n, re_maps, direction):

    """evaluate if there is support from neighboring tiles, plot landscape of the top_n, return candidate list"""
    val = 0
    already_run = []

    remove_tiles = []
    # get corresponding enhancer id
    for tile_id in candidate_list:
        enh_id = getEnhid(tile_id, df)
        
        #multiple tiles per enhancer
        if enh_id not in already_run and enh_id is not None:
            
            # get enh_id for tile coordinate
            enh_id = getEnhid(tile_id, enh)
        
            cols = ["sig_n_dif_n_bs", "glut_1", "glut_2", "glut_3", "gaba_1",
                    "gaba_2", "gaba_3", "coor", "enh.id", "enh.name", "tile.order", "gene"]
        
            t = df.loc[df["enh.id"] == enh_id].drop_duplicates().copy()
        
            # stratify by gene name, as one enhancer may be near two different genes
            enh_name = list(set(t["enh.name"]))[0]
            gene = getMinDist(t)
        
            # get support (number of significant calls) per tile in enhancer
            support = getSupport(enh_id, t)
    
            # scan neighbors for supporting significant differences in activity (at least 3/5 neighbor tiles must have support). 
            neighbor_support = supportingNeighbors(support)
    
            # if there are significant neighborhoods, then count this enhancer. 
            if len(neighbor_support) > 0:
    
                # count the enhancer, so that you don't run the same enhancer twice, even when there may be multiple significant tiles. 
                already_run.append(enh_id)
                val+=1
    
                # plot
                if val < top_n:
                        
                    t = t[cols[:-1]].fillna(0).drop_duplicates()  # subset the dataframe to just these columns
                    
                    # if the outpath does not exist, make it
                    if os.path.exists(os.path.join(re_maps, direction)) is False:
                        os.mkdir(os.path.join(re_maps, direction))
                
                   
                    # put enhancer data into long form, plot tiles
                    plot_support_annot = "support_name"
                    
                    # make an output file name for tile plot.
                    out_line = os.path.join(re_maps, direction,
                                            f"{gene}.{enh_name}.line.{plot_support_annot}.pdf")
                
                    # get long form tile activity across replicates for gaba, glut.
                    longform_gaba, var_name = getLongFormEnhInfo(t, 'gaba')
                    longform_glut, var_name = getLongFormEnhInfo(t, 'glut')
            
                    # measure neighbor activity
                    tile_activity = neighborActivity(longform_glut, longform_gaba, neighbor_support)
                    
                    # plot line
                    #plot_lineplot(longform_gaba, longform_glut, var_name, out_line, enh_id,  enh_name, gene, support, plot_support_annot)
            
                    # plot the name of the support
                    #plot_support_annot = "support"
                    
                    # make an output file name for tile plot.
                    #out_line = os.path.join(re_maps, direction,
                                          #  f"{gene}.{enh_name}.line.{plot_support_annot}.pdf")
                    
                    # plot line
                    # plot_lineplot(longform_gaba, longform_glut, var_name, out_line, enh_id,  enh_name, gene, support, plot_support_annot)
            else:
                print("remove this tile", tile_id)
                remove_tiles.append(tile_id)
           
    return remove_tiles
        

## Top 8 GABA candidates

ACTIVE_CL = "GABA"
INACTIVE_CL = "GLUT"
direction = f"positive {ACTIVE_CL}>negative {INACTIVE_CL}"

# get significant tiles that fall in direction
SET = getSigTileIdDirection(direction, enh, 2)

# subset dataframe and sort
candidate_cols = ["name", ACTIVE_CL.lower(), 'delta.gaba-glut', "tile.order"]

# order candidates by highest activity, greatest difference
GABA_candidates = enh.loc[enh["name"].isin(SET), candidate_cols].drop_duplicates().sort_values(
    by=[ACTIVE_CL.lower(), 'delta.gaba-glut'], ascending=False)["name"].to_list()


print(len(GABA_candidates))

# hand-picked GLUT candidates

candidate_genes = ["OPRM1", "ELMO1", "NBPF1", "SLC2A1", "EPHB2", "SYNGAP1", "SLC6A1", "JAG2"]

#tile_id = GLUT_HANDPICK[0]
df = enh.copy()

candidate_dict={}

"""evaluate if there is support from neighboring tiles in an enhancer, plot landscape of the top_n, return candidate list"""

candidate_enh_gene = df.loc[df["gene"].isin(candidate_genes), ["gene", "enh.id", "enh.name"]].drop_duplicates()

for enh_id in candidate_enh_gene["enh.id"].unique():
    print(enh_id)
    
    #multiple tiles per enhancer
    #if enh_id not in already_run and enh_id is not None:
    
    cols = ['#chr', 'start.tile', 'end.tile',
            "sig_n_dif_n_bs", "glut_1", "glut_2", "glut_3", 
            "gaba_1", "gaba_2", "gaba_3",
            "gaba", "glut", "coor", "enh.id", "enh.name",
            "tile.order", "direction"]
    
    t = df.loc[df["enh.id"] == enh_id].drop_duplicates().copy()
    
    # stratify by gene name, as one enhancer may be near two different genes
    enh_name = list(set(t["enh.name"]))[0]
    gene = getMinDist(t)
    
    # get support (number of significant calls) per tile in enhancer
    support = getSupport(enh_id, t)
    
    # scan neighbors for supporting significant differences in activity (at least 3/5 neighbor tiles must have support). 
    neighbor_support, sig_neighbor_tiles = supportingNeighbors(support)
    sig_neighbor_tiles =list(sig_neighbor_tiles)
    t = t[cols].fillna(0).drop_duplicates()  # subset the dataframe to just these columns

    # add annotations
    t["gene"] = gene
    t["candidate"] = False
    t["candidate_list"] = None
    if neighbor_support!=None:

        # if there are gaps between significant tiles, fill those gabs in!
        filled_in_neighbors = fillInSigGaps(sig_neighbor_tiles)

        # hard coded because this is a tricky tile w/ flanks missing data. 
        if enh_name =="enh.43":  # nbpf1
            filled_in_neighbors.append([25, 26, 27]) 
        elif enh_name =="enh.123":  # slc2a1
            filled_in_neighbors = [list(np.arange(66,79))]
            print("FILLIN", filled_in_neighbors)
        elif enh_name =='enh.1159':
            filled_in_neighbors = [list(np.arange(1,7))]
        elif enh_name =='enh.1157':
            filled_in_neighbors = [list(np.arange(1,5))]
        elif enh_name =="enh.1154":
            filled_in_neighbors = [list(np.arange(12,18))]
        # annotate dataframe with significant regions
        for i in filled_in_neighbors:
            t.loc[t["tile.order"].isin(i), "candidate"] =True
            t.loc[t["tile.order"].isin(i), "candidate_list"] = ",".join(str(n) for n in i)
            print(enh_id, ",".join(str(n) for n in i))
    
        candidate_dict[enh_id] = t

    

candidates = pd.concat(candidate_dict.values())

candidates.loc[candidates["enh.name"]=="enh.43", "candidate_list"].unique()

## get direction, start, end of candidate regions summarized across significant tiles, with gaps filled in. 

candidates["direction"] = candidates["direction"].astype(str)

directions = candidates.loc[candidates["candidate"]==True].groupby(
    ["enh.name","gene", 'enh.id', "candidate_list"])["direction"].sum().reset_index().sort_values(by="enh.name")

starts = candidates.loc[candidates["candidate"]==True].groupby(
    ["enh.name","gene", 'enh.id', "candidate_list"])["start.tile"].min().reset_index().sort_values(by="enh.name")

ends = candidates.loc[candidates["candidate"]==True].groupby(
    ["enh.name","gene", 'enh.id', "candidate_list"])["end.tile"].max().reset_index().sort_values(by="enh.name")

### add summarized information back together

summed = pd.merge(starts, ends) # combine min start, max end of sig tile runs
summed = pd.merge(directions, summed) # add direction annotations for sig tile runs
summed = pd.merge(summed, df[["enh.id", "#chr"]], how="left").drop_duplicates()# add chr annotations
summed.shape

## reorg cols and write.

summed_cols = [ '#chr', 
                'start.tile',
                'end.tile',
                'enh.name',
                'gene',
                'enh.id',
                'candidate_list',
                'direction',
]

candidates = pd.merge(candidates, summed[["enh.name", "candidate_list"]], how="left")
candidates.drop_duplicates().to_csv(CANDIDATE_REGIONS, sep='\t', index=False)
summed[summed_cols].drop_duplicates().to_csv(SUMMED_CANDIDATE_REGIONS, sep='\t', index=False)

summed.groupby(["gene", "enh.name"])['candidate_list'].unique()

# Candidates

enh_tiles = {"enh.1154":[12,13,14,15,16,17], 
            "enh.1159":[1,2,3,4,5,6], 
             "enh.1157":[1,2,3,4], 
             "enh.74":[18,19,20,21,22,23],
             "enh.43":[25,26,27],
             "enh.1531":[0,1,2,3,4,5],
             "enh.1503":[0,1,2,3,4,5,6,7,8,9,10,11,12,13],
             "enh.1628":[0,1,2,3,4,5,6,7]
             
            }

GABA = ["enh.1154", "enh.1159", "enh.1157", "enh.74", "enh.43"]
GLUT = ["enh.1531", "enh.1603", "enh.1628"]


# merged sequences. 
# highlight best single tile. 


list(enh)

def getEnhDf(enh_name, tile_list, enh_df):
    """ get enhancer only dataframe, sort by tile, return dataframe subset"""
    cols = ["sequence", "enh.name", "coor", 'enh.id',
            "tile.order", 'gaba',
         'glut', 'delta.gaba-glut',
            'celltype_dif', "fdr_bool", "bs", 'direction']
    
    sub_df = enh_df.loc[(enh_df["enh.name"]==enh_name)&
                    (enh_df["tile.order"].isin(tile_list)),
                    cols].drop_duplicates().sort_values(by="tile.order").reset_index(drop=True)

    return sub_df


def mergeSeq(enh_name, tile_list, enh_df, length_merged):
    """merge sequences together for consecutive candidate tiles"""
    
    last_seq, full_seq = "", ""  # record the last sequence snippet and the full sequence snippet
    
    for n, i in enh_df.iterrows():
    
        seq = i['sequence'][15:-15]  ## remove adaptor sequences ( first and last 15 bp)
        if last_seq == "":
            last_seq, full_seq = seq, seq
            
        elif last_seq[20:] == seq[:-20]:  # check that the internal sequences match
            full_seq+=seq[-20:] # append the last 20 bases
            last_seq = seq
            
    if len(full_seq) == length_merged:
        print("complete merge")

    return full_seq


def labelSigTile(test):
    """label the best candidate tile per sequence"""
    # sum the number of significant calls per tiles
    sig = test[test.columns[-4:-1]].sum(axis=1).copy()  
    
    # get max significant counts in element
    most_sig_calls = sig.max()  
    
    # find tile indexes with max significant calls 
    sig_indexes = sig.loc[sig== most_sig_calls].index 
    
    # filter on direction, get coordinate. Require that direction is negative for one element an positive for another. 
    if test["enh.name"].iloc[0] in GABA:
        direction ="positive GABA>negative GLUT"
    else:
        direction ="positive GLUT>negative GABA"
        
    best_tile_coor = list(test.loc[(test["direction"]==direction)&
                              (test.index.isin(sig_indexes)), "coor"])
    # annotate best_tile
    test["best_tile"] = ""
    
    if len(best_tile_coor)>1:
        """find biggest difference tile, label that oen the best"""
        biggest_dif = test.loc[test["coor"].isin(best_tile_coor), 'delta.gaba-glut'].abs().max() 
        
        # label the most sig and biggest difference
        test.loc[(test["coor"].isin(best_tile_coor))&
        (test['delta.gaba-glut'].abs()==biggest_dif),
        "best_tile"] = "best"
    else:
        """find the most significant tile, name that one the best"""
        test.loc[test["coor"].isin(best_tile_coor), "best_tile"] = "best"

    return test


new_sum , new_tiles= {}, {}
for key, vals in enh_tiles.items():
    val_str=",".join([str(i) for i in vals])

    # get the summarized data
    summed_key = summed.loc[(summed["enh.name"]==key)&
              (summed['candidate_list']==val_str)].drop_duplicates()
    length = 270 + ((len(vals)-1)*20) # minus one corrects for the lack of 20bp stride in the first tile
    length_merged = summed_key["end.tile"]-summed_key["start.tile"]  

    # make sure the tiles are continuous
    #if length == length_merged.iloc[0]:

    # get enh df
    candidate_df = getEnhDf(key, vals, enh)

    # get merged sequence
    merged_sequence = mergeSeq(enh_name, vals, candidate_df, length)

    # label best candidate
    labeled = labelSigTile(candidate_df)

    if key == "enh.1531":
        labeled.loc[labeled["tile.order"] == 4, 'best_tile'] = "best"
        print("labeled")
        
    new_tiles[key] = labeled

    # add merged sequence
    summed_key["sequence"] = merged_sequence
    summed_key["sequence_len"] = length
    new_sum[key] = summed_key
    print(key)

### make best tile, all candidate tiles, merged tile files with sequences. 

new_sumdf = pd.concat(new_sum.values())  # summaried candidate regions
new_tiledf = pd.concat(new_tiles.values())  # all candidate tiles

#### rename columns for legibility

new_sumdf.columns =['enh.name',
                     'gene',
                     'enh.id',
                     'candidate.list',
                     'direction',
                     'start.tile.merged',
                     'end.tile.merged',
                     '#chr',
                     'sequence.no.adapters',
                     'sequence.len']
new_tiledf.columns = ['sequence.no.adapters',
                     'enh.name',
                     'tile.coor',
                     'enh.id',
                     'tile.order',
                     'gaba.log2meanZ',
                     'glut.log2meanZ',
                     'delta.gaba-glut',
                     'celltype.dif',
                     'fdr.bool',
                     'bs',
                     'direction',
                     'best.tile']
new_tiledf["sequence.len"] = new_tiledf["sequence.no.adapters"].apply(lambda x: len(x))

new_tiledf = new_tiledf[['enh.name',
                     'tile.coor',
                     'enh.id',
                     'tile.order',
                     'gaba.log2meanZ',
                     'glut.log2meanZ',
                     'delta.gaba-glut',
                     'celltype.dif',
                     'fdr.bool',
                     'bs',
                     'direction',
                     'best.tile',
                         "sequence.len", 
                         'sequence.no.adapters']]

new_tiledf["sequence.no.adapters"].apply(lambda x:x[:15])

# remove adaptors
new_tiledf["sequence.no.adapters"] = new_tiledf["sequence.no.adapters"].apply(lambda x:x[15:-15])

best = new_tiledf.loc[new_tiledf["best.tile"]!=""]  # all best candidate tiles

### visual inspection

### summarized regions

new_sumdf

### all tiles

new_tiledf.loc[new_tiledf["enh.name"]=="enh.1531"]

### best

best

## write

out = os.path.join(RE, "SUMMARIZED.CANDIDATE.TILE.SEQUENCES.tsv")
new_sumdf.to_csv(out, sep="\t", index=False)

out = os.path.join(RE, "BEST.CANDIDATE.TILE.SEQUENCES.tsv")
best.to_csv(out, sep="\t", index=False)

out = os.path.join(RE, "ALL.CANDIDATE.TILE.SEQUENCES.tsv")
new_tiledf.to_csv(out, sep="\t", index=False)
